Renamed test utils
