+++ packages/gafl-webapp-service/src/__tests__/server-test-sanitize.js
@@ -1,4 +1,4 @@
-import { start, stop, initialize, injectWithCookies, server } from '../__mocks__/test-utils.js'
+import { start, stop, initialize, injectWithCookies, server } from '../__mocks__/test-utils-system.js'
 
 beforeAll(d => {
   start(d)
