+++ packages/gafl-webapp-service/src/__tests__/server.telesales.spec.js
@@ -1,4 +1,4 @@
-import { start, stop, initialize, injectWithCookies } from '../__mocks__/test-utils.js'
+import { start, stop, initialize, injectWithCookies } from '../__mocks__/test-utils-system.js'
 import { LICENCE_LENGTH } from '../uri.js'
 
 beforeAll(d => {
