+++ packages/gafl-webapp-service/src/handlers/__tests__/controller-handler.spec.js
@@ -1,4 +1,4 @@
-import { start, stop, initialize, injectWithCookies } from '../../__mocks__/test-utils.js'
+import { start, stop, initialize, injectWithCookies } from '../../__mocks__/test-utils-system.js'
 import { CONTROLLER, DATE_OF_BIRTH, ADD_PERMISSION, NEW_TRANSACTION } from '../../uri.js'
 
 beforeAll(d => start(d))
