+++ packages/gafl-webapp-service/src/pages/contact/newsletter/__tests__/newsletter.spec.js
@@ -11,9 +11,9 @@ import {
 
 import { HOW_CONTACTED } from '../../../../processors/mapping-constants.js'
 
-import { start, stop, initialize, injectWithCookies, postRedirectGet } from '../../../../__mocks__/test-utils.js'
+import { start, stop, initialize, injectWithCookies, postRedirectGet } from '../../../../__mocks__/test-utils-system.js'
 
-import { ADULT_TODAY, dobHelper } from '../../../../__mocks__/test-helpers'
+import { ADULT_TODAY, dobHelper } from '../../../../__mocks__/test-utils-business-rules'
 import { licenceToStart } from '../../../licence-details/licence-to-start/update-transaction'
 
 beforeAll(d => start(d))
