+++ packages/gafl-webapp-service/src/pages/payment/failed/__tests__/failed-spec.js
@@ -1,4 +1,4 @@
-import { initialize, injectWithCookies, start, stop } from '../../../../__mocks__/test-utils.js'
+import { initialize, injectWithCookies, start, stop } from '../../../../__mocks__/test-utils-system.js'
 import { PAYMENT_FAILED } from '../../../../uri.js'
 
 beforeAll(d => start(d))
